added VehiclePOI flow and filters

+++ src/main/java/com/izabel/localiza/domain/POI.java
@@ -23,11 +23,11 @@ public class POI {
     private String name;
 
     @Column
-    private Float radius;
+    private Double radius;
 
     @Column
-    private Float latitude;
+    private Double latitude;
 
     @Column
-    private Float longitude;
+    private Double longitude;
 }
